Patch recline_view.js to avoid race condition and prevent error
When this javascript module is loaded in an iframe it often (particularly
with ckan in debug mode) will fail as the ckan.views.filters property hasn't
been set yet. This is because of a race condition with the parent window
which is loading the javascript which sets this in parallel and just hasn't
finished loading and running it yet.

To counteract this this patch adds a wait in to make sure the properties we
need are ready. After 5 seconds of checking every 50ms it just carries on
anyway (likely resulting in an error) to avoid an infinite loop.