Fix attribute error in reclineView.js patch
self.options.i18n.noRecords fails because i18n is undefined. This
isn't an upgrade issue, it is failing on main too. Looks like we're
just not passing this option through to the recline module (see our
list view for an example of where we do do this). This should
probably be i18n'd but meh, it's not the biggest issue given that
this is our fork of ckan that like no one will ever use but us. If
in the future we properly support other languages we'll just have
to fix it then but hopefully whenever that day comes we won't be
using any of the views?